Shortened the parse tree; Correcting fallout.
Removed node generation for many productions that only passed their
children up the tree. Unfortunately, I happened to rely on a lot of
those, especially implicitly. Had to correct some functions so they
could deal with the node they're looking for being the node they're
looking at -- because that node got pushed up to the top of a subtree.